+++ src/routes/+layout.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import "../app.css";
 	import { onMount } from "svelte";
+	import HokoTracking from "$lib/components/site/shell/HokoTracking.svelte";
 	import SiteShell from "$lib/components/site/shell/SiteShell.svelte";
 	import type { LayoutProps } from "./$types";
 
@@ -11,6 +12,10 @@
 		performance.mark("lab-hydrated");
 	});
 </script>
+
+{#if data.publication.name === "canonical"}
+	<HokoTracking />
+{/if}
 
 <SiteShell
 	publicationProfile={data.publication.name}